[ll] Move command pools into a module, implement vulkan dummy backend and fix SubpassCommandBuffer constraint
